+++ backend/fcm-server/src/main/java/news/bombom/notification/domain/ArticleArrivalNotification.java
@@ -2,33 +2,20 @@ package news.bombom.notification.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import java.time.LocalDateTime;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
-import news.bombom.notification.common.BaseEntity;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ArticleArrivalNotification extends BaseEntity {
+public class ArticleArrivalNotification extends Notification {
 
     private static final int MAX_RETRY_ATTEMPTS = 6;
     private static final int SHORT_RETRY_ATTEMPTS = 3;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private Long memberId;
 
     @Column(nullable = false)
     private Long articleId;
@@ -38,18 +25,6 @@ public class ArticleArrivalNotification extends BaseEntity {
 
     @Column(nullable = false)
     private String articleTitle;
-
-    @Column(nullable = false)
-    @Enumerated(EnumType.STRING)
-    private NotificationStatus status;
-
-    @Column(nullable = false, columnDefinition = "INT DEFAULT 0")
-    private int attempts = 0;
-
-    private LocalDateTime nextRetryAt;
-
-    @Column(length = 1024)
-    private String lastError;
 
     @Column(columnDefinition = "BOOLEAN DEFAULT FALSE")
     private boolean isRead = false;
@@ -67,36 +42,21 @@ public class ArticleArrivalNotification extends BaseEntity {
             String lastError,
             boolean isRead
     ) {
+        super(memberId, status, attempts, nextRetryAt, lastError);
         this.id = id;
-        this.memberId = memberId;
         this.articleId = articleId;
         this.newsletterName = newsletterName;
         this.articleTitle = articleTitle;
-        this.status = status != null ? status : NotificationStatus.PENDING;
-        this.attempts = attempts;
-        this.nextRetryAt = nextRetryAt;
-        this.lastError = lastError;
         this.isRead = isRead;
     }
 
-    public void markSent() {
-        this.status = NotificationStatus.SENT;
-        this.nextRetryAt = null;
-        this.lastError = null;
-    }
-
-    public void markFailed(String reason) {
-        this.status = NotificationStatus.FAILED;
-        this.attempts++;
-        this.lastError = reason;
-        this.nextRetryAt = calculateNextRetryTime(this.attempts);
-    }
-
+    @Override
     public boolean shouldRetry() {
         return this.attempts <= MAX_RETRY_ATTEMPTS;
     }
 
-    private LocalDateTime calculateNextRetryTime(int attempts) {
+    @Override
+    public LocalDateTime calculateNextRetryTime(int attempts) {
         long delaySeconds;
 
         if (attempts <= SHORT_RETRY_ATTEMPTS) {
